Decouple code
